Fix: Firebase app singleton in tests affected
