+++ src/utils.py
@@ -8,3 +8,14 @@ def skip_none_dict_factory(d: list[tuple[str, Any]]) -> dict:
 
 def as_dict_skip_none(d) -> dict:
     return asdict(d, dict_factory=skip_none_dict_factory)
+
+
+def singleton(cls):
+    instances = {}
+
+    def getinstance(*args, **kwargs):
+        if cls not in instances:
+            instances[cls] = cls(*args, **kwargs)
+        return instances[cls]
+
+    return getinstance
